backend/pyproject.toml
[project]
name = "fake-currency-backend"
version = "1.0.0"
description = "LeJEPA-based fake Indian currency detection backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "cryptography>=44.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "timm>=1.0.0",
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.22.0",
    "scikit-learn>=1.4.0",
    "easyocr>=1.7.0",
    "numpy>=1.26.0",
    "aiofiles>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]
